Creating the content in files

+++ service-a/app/schemas.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 
 class Coordinats(BaseModel):
     ip: str
-    lat: int
-    lon: int
+    lat: float
+    lon: float
